requirements.txt
spacy
flask
flask-sqlalchemy
pymysql
